Make pendulum animation robust to wraparound.
Previously the animation stored the midpoint of the swing. This goes kooky
when an entity wraps around the map, because the acceleration jumps way too high.
To solve this, I changed the animation to store the total distance from the pendulum,
incrementing it after an entity moves. This is robust even to multiple wraparounds
(given a pendulum with very high velocity).