Clean up GameClient.
GameClient is split into 2 structs, one for polling the servert and one for pushing inputs to the server.
The GameClient previously on the main thread is now just an UnboundedSender.

A few different functionalities were conflated by GameClient; it should be clearer now what is responsible for what.

Additionally, only one runtime is started now; previously a runtime was created just to pull the initial game state
from the server, after which it was discarded. Now, in lieu of a one-off runtime, the main thread uses a
condition variable to wait until the runtime has initialized the game state.