Handle race in unregistration.
Unregistration drop fn removes the game from the games map, but that
can run after a new version of the game has been inserted and the health
check for the previous version was aborted. This is fixed by adding a
version number to the game state, which is checked before unregistering.